refactor(apple): remove UI code from model
- Move NSAlert creation from SessionNotification model to MacOSAlert view
- Remove unnecessary SwiftUI import from SettingsViewModel (only needs Combine)

Models should not depend on AppKit/UIKit.
