+++ swift/apple/FirezoneKit/Sources/FirezoneKit/ViewModels/SettingsViewModel.swift
@@ -6,7 +6,6 @@
 
 import Combine
 import Foundation
-import SwiftUI
 
 @MainActor
 class SettingsViewModel: ObservableObject {
